requirements.txt
urwid==2.0.1
nose-cov
nose
flake8
PyYAML
testtools
mock
attrs
systemd-python
jsonschema
pyudev
requests
requests-unixsocket
aiohttp
-e git+https://github.com/CanonicalLtd/probert@b697ab779e7e056301e779f4708a9f1ce51b0027#egg=probert
